test: separate test case for LogMessageDirector
